##
The most significant changes in the code involve the update of the assembly version and the refactoring of the method for pushing a package to the NuGet server. The assembly version has been updated from `1.1.8875.3227` to `1.1.8875.7250` in both `AssemblyInfo.cs` and `PublishPackageToNuGet2017.csproj` files, indicating that a new version of the assembly has been prepared. The method for pushing a package to the NuGet server has been refactored in the `NuGetPkgPublishService.cs` file. The previous code has been replaced with a new method `PushNuGetPackage` that uses the `dotnet` command-line interface, providing a more direct and possibly more reliable method of pushing the package.

List of changes:

1. The `AssemblyVersion`, `AssemblyFileVersion`, and `AssemblyInformationalVersion` have been updated in `AssemblyInfo.cs` and `PublishPackageToNuGet2017.csproj` files. (Refer to the changes in these files)
2. Two new namespaces, `System.Diagnostics` and `System.Threading.Tasks`, have been added to the `NuGetPkgPublishService.cs` file. (Refer to the changes in `NuGetPkgPublishService.cs`)
3. The method of pushing a package to the NuGet server has been refactored in the `NuGetPkgPublishService.cs` file. The previous code has been replaced with a new method `PushNuGetPackage`. (Refer to the changes in `NuGetPkgPublishService.cs`)
4. A new method `PushNuGetPackage` has been added to the `NuGetPkgPublishService.cs` file. This method uses the `dotnet` command-line interface to push a NuGet package. (Refer to the changes in `NuGetPkgPublishService.cs`)